# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "vibe80 REST API",
    "version": "1.0.0",
    "description": "Spec (workspace auth, providers at workspace level)"
  },
  "servers": [
    {
      "url": "http://localhost:5179"
    }
  ],
  "paths": {
...
    "/api/workspaces": {
      "post": {
        "summary": "Create workspace",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkspaceCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Workspace created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceCreateResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        },
        "security": []
      }
    },
    "/api/workspaces/login": {
      "post": {
        "summary": "Login workspace (exchange secret for token)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkspaceLoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Workspace token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceLoginResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        },
        "security": []
      }
    },
    "/api/workspaces/refresh": {
      "post": {
        "summary": "Refresh workspace token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkspaceRefreshRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Workspace token refreshed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceRefreshResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        },
        "security": []
      }
    },
    "/api/workspaces/{workspaceId}": {
      "delete": {
        "summary": "Delete workspace (policy TBD)",
        "parameters": [
          {
            "$ref": "#/components/parameters/WorkspaceId"
          }
        ],
        "responses": {
          "200": {
            "description": "Workspace deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      },
      "patch": {
        "summary": "Update workspace providers/auth",
        "parameters": [
          {
            "$ref": "#/components/parameters/WorkspaceId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkspaceUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Workspace updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceUpdateResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
...
      "WorkspaceId": {
        "name": "workspaceId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "pattern": "^w[0-9a-f]{24}$"
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Bad request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
      "Forbidden": {
        "description": "Forbidden",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
      "ServerError": {
        "description": "Server error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    },
    "schemas": {
...
      "OkResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          }
        },
        "required": [
          "ok"
        ],
        "additionalProperties": true
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": true
      },
...
      "WorkspaceProviderAuth": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "api_key",
              "auth_json_b64",
              "setup_token"
            ]
          },
          "value": {
            "type": "string"
          }
        },
        "required": [
          "type"
        ],
        "additionalProperties": false
      },
      "WorkspaceProviderConfig": {
        "type": "object",
        "properties": {
          "enabled": {
            "type": "boolean"
          },
          "auth": {
            "$ref": "#/components/schemas/WorkspaceProviderAuth"
          }
        },
        "additionalProperties": false
      },
      "WorkspaceCreateRequest": {
        "type": "object",
        "properties": {
          "providers": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/WorkspaceProviderConfig"
            }
          }
        },
        "required": [
          "providers"
        ],
        "additionalProperties": false
      },
      "WorkspaceCreateResponse": {
        "type": "object",
        "properties": {
          "workspaceId": {
            "type": "string"
          },
          "workspaceSecret": {
            "type": "string"
          }
        },
        "required": [
          "workspaceId",
          "workspaceSecret"
        ],
        "additionalProperties": false
      },
      "WorkspaceLoginRequest": {
        "type": "object",
        "properties": {
          "workspaceId": {
            "type": "string"
          },
          "workspaceSecret": {
            "type": "string"
          }
        },
        "required": [
          "workspaceId",
          "workspaceSecret"
        ],
        "additionalProperties": false
      },
      "WorkspaceLoginResponse": {
        "type": "object",
        "properties": {
          "workspaceToken": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          },
          "expiresIn": {
            "type": "integer",
            "description": "Seconds until expiration (1h)"
          },
          "refreshExpiresIn": {
            "type": "integer",
            "description": "Seconds until refresh expiration (30d)"
          }
        },
        "required": [
          "workspaceToken"
        ],
        "additionalProperties": true
      },
      "WorkspaceRefreshRequest": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string"
          }
        },
        "required": [
          "refreshToken"
        ],
        "additionalProperties": false
      },
      "WorkspaceRefreshResponse": {
        "type": "object",
        "properties": {
          "workspaceToken": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          },
          "expiresIn": {
            "type": "integer",
            "description": "Seconds until expiration (1h)"
          },
          "refreshExpiresIn": {
            "type": "integer",
            "description": "Seconds until refresh expiration (30d)"
          }
        },
        "required": [
          "workspaceToken",
          "refreshToken"
        ],
        "additionalProperties": true
      },
      "WorkspaceUpdateRequest": {
        "type": "object",
        "properties": {
          "providers": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/WorkspaceProviderConfig"
            }
          }
        },
        "required": [
          "providers"
        ],
        "additionalProperties": false
      },
      "WorkspaceUpdateResponse": {
        "type": "object",
        "properties": {
          "workspaceId": {
            "type": "string"
          },
          "providers": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/WorkspaceProviderConfig"
            }
          }
        },
        "required": [
          "workspaceId",
          "providers"
        ],
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ]
}